Move QueryParamsEncoder.kt from "utils" to "jitm" package.
